+++ packages/utilities/fast-web-utilities/src/dom.spec.ts
@@ -29,7 +29,8 @@ describe("isHTMLElement", () => {
         expect(isHTMLElement(document.getElementById("element"))).to.equal(true);
     });
     it("should return false if all arguments are NOT HTML elements", () => {
-        expect(isHTMLElement(document.getElementById("element").childNodes)).to.equal(
+        /* eslint-disable-next-line @typescript-eslint/no-non-null-assertion */
+        expect(isHTMLElement(document.getElementById("element")!.childNodes)).to.equal(
             false
         );
     });