+++ packages/utilities/fast-web-utilities/src/dom.ts
@@ -11,10 +11,10 @@ export function isHTMLElement(...args: any[]): boolean {
  * Returns all displayed elements inside of a root node that match a provided selector
  */
 export function getDisplayedNodes(
-    rootNode: HTMLElement,
-    selector: string
+    rootNode: HTMLElement | null | undefined,
+    selector: string | null | undefined
 ): HTMLElement[] | void {
-    if (!isHTMLElement(rootNode)) {
+    if (!rootNode || !selector || !isHTMLElement(rootNode)) {
         return;
     }
 
@@ -29,7 +29,7 @@ export function getDisplayedNodes(
  * Gets the numeric key code associated with a keyboard event. This method is for use with DOM level 3 events
  * that still use the deprecated keyCode property.
  */
-export function getKeyCode(event: KeyboardEvent): number {
+export function getKeyCode(event: KeyboardEvent | null): number | null {
     return event === null ? null : event.which || event.keyCode || event.charCode;
 }
 
@@ -50,7 +50,7 @@ function getNonce(): string | null {
 /**
  * Test if the document supports :focus-visible
  */
-let _canUseFocusVisible: boolean;
+let _canUseFocusVisible: boolean | undefined;
 export function canUseFocusVisible(): boolean {
     if (typeof _canUseFocusVisible === "boolean") {
         return _canUseFocusVisible;
@@ -85,7 +85,7 @@ export function canUseFocusVisible(): boolean {
     return _canUseFocusVisible as boolean;
 }
 
-let _canUseCssGrid: boolean;
+let _canUseCssGrid: boolean | undefined;
 export function canUseCssGrid(): boolean {
     if (typeof _canUseCssGrid === "boolean") {
         return _canUseCssGrid;