+++ packages/web-components/fast-foundation/src/foundation-element/foundation-element.ts
@@ -208,6 +208,7 @@ function resolveOption<T, K extends FoundationElementDefinition>(
  *
  * @internal
  */
+/* eslint-disable @typescript-eslint/no-unused-vars */
 export class FoundationElementRegistry<
     TDefinition extends FoundationElementDefinition,
     TType
@@ -284,3 +285,4 @@ export class FoundationElementRegistry<
         });
     }
 }
+/* eslint-enable @typescript-eslint/no-unused-vars */