+++ packages/web-components/fast-foundation/src/picker/picker.ts
@@ -500,6 +500,7 @@ export class Picker extends FormAssociatedPicker {
 
         this.inputElement.addEventListener("input", this.handleTextInput);
         this.inputElement.addEventListener("click", this.handleInputClick);
+        /* eslint-disable-next-line @typescript-eslint/no-non-null-assertion */
         this.$fastController.addBehaviors([this.itemsRepeatBehavior!]);
 
         this.menuElement.suggestionsAvailableText = this.suggestionsAvailableText;
@@ -514,6 +515,7 @@ export class Picker extends FormAssociatedPicker {
             { positioning: true }
         ).createBehavior(this.optionsPlaceholder);
 
+        /* eslint-disable-next-line @typescript-eslint/no-non-null-assertion */
         this.$fastController.addBehaviors([this.optionsRepeatBehavior!]);
 
         this.handleSelectionChange();
@@ -830,9 +832,10 @@ export class Picker extends FormAssociatedPicker {
                     e.target.value
                 }`;
             }
-            this.toggleFlyout(false);
             this.inputElement.value = "";
             this.query = "";
+            this.inputElement.focus();
+            this.toggleFlyout(false);
             return false;
         }