+++ packages/web-components/fast-foundation/src/switch/switch.ts
@@ -45,102 +45,15 @@ export class Switch extends FormAssociatedSwitch {
     public initialValue: string = "on";
 
     /**
-     * The checked attribute value. This sets the initial checked value.
-     *
-     * @public
-     * HTML Attribute: checked
-     */
-    @attr({ attribute: "checked", mode: "boolean" })
-    public checkedAttribute: boolean;
-    private checkedAttributeChanged(): void {
-        this.defaultChecked = this.checkedAttribute;
-    }
-
-    /**
      * @internal
      */
     @observable
     public defaultSlottedNodes: Node[];
 
-    /**
-     * Initialized to the value of the checked attribute. Can be changed independently of the "checked" attribute,
-     * but changing the "checked" attribute always additionally sets this value.
-     *
-     * @public
-     */
-    @observable
-    public defaultChecked: boolean;
-    private defaultCheckedChanged(): void {
-        if (!this.dirtyChecked) {
-            // Setting this.checked will cause us to enter a dirty state,
-            // but if we are clean when defaultChecked is changed, we want to stay
-            // in a clean state, so reset this.dirtyChecked
-            this.checked = this.defaultChecked;
-            this.dirtyChecked = false;
-        }
-    }
-
-    /**
-     * The checked state of the control.
-     *
-     * @public
-     */
-    @observable
-    public checked: boolean;
-    private checkedChanged(): void {
-        if (!this.dirtyChecked) {
-            this.dirtyChecked = true;
-        }
-
-        this.updateForm();
-
-        if (this.proxy instanceof HTMLInputElement) {
-            this.proxy.checked = this.checked;
-        }
-
-        this.$emit("change");
-
-        this.checked ? this.classList.add("checked") : this.classList.remove("checked");
-
-        this.validate();
-    }
-
-    /**
-     * Tracks whether the "checked" property has been changed.
-     * This is necessary to provide consistent behavior with
-     * normal input checkboxes
-     */
-    private dirtyChecked: boolean = false;
-
-    /**
-     * @internal
-     */
-    public connectedCallback(): void {
-        super.connectedCallback();
-
-        this.proxy.setAttribute("type", "checkbox");
-
-        this.updateForm();
-    }
-
     public constructor() {
         super();
 
-        this.defaultChecked = !!this.checkedAttribute;
-        this.checked = this.defaultChecked;
-    }
-
-    /**
-     * @internal
-     */
-    public formResetCallback = (): void => {
-        this.checked = this.checkedAttribute;
-        this.dirtyChecked = false;
-    };
-
-    private updateForm(): void {
-        const value = this.checked ? this.value : null;
-        this.setFormValue(value, value);
+        this.proxy.setAttribute("type", "checkbox");
     }
 
     /**
@@ -163,4 +76,15 @@ export class Switch extends FormAssociatedSwitch {
             this.checked = !this.checked;
         }
     };
+
+    /**
+     * @internal
+     */
+    public checkedChanged(prev: boolean | undefined, next: boolean) {
+        super.checkedChanged(prev, next);
+        /**
+         * @deprecated - this behavior already exists in the template and should not exist in the class.
+         */
+        this.checked ? this.classList.add("checked") : this.classList.remove("checked");
+    }
 }