+++ sites/fast-component-explorer/webpack.dev.js
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable no-undef */
 const merge = require("webpack-merge");
 const baseConfig = require("./webpack.common");