+++ packages/utilities/fast-react-wrapper/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @microsoft/fast-react-wrapper
 
-This log was last generated on Fri, 22 Oct 2021 17:40:52 GMT and should not be manually modified.
+This log was last generated on Sun, 31 Oct 2021 07:17:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.10
+
+Sun, 31 Oct 2021 07:17:45 GMT
+
+### Patches
+
+- update fast eslint package version (chhol@microsoft.com)
+
+## 0.1.9
+
+Thu, 28 Oct 2021 07:15:32 GMT
+
+### Patches
+
+- Bump @microsoft/fast-foundation to v2.23.2 (nicholasrice@users.noreply.github.com)
 
 ## 0.1.8