+++ sites/fast-color-explorer/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @microsoft/fast-color-explorer
 
-This log was last generated on Fri, 22 Oct 2021 17:40:52 GMT and should not be manually modified.
+This log was last generated on Sun, 31 Oct 2021 07:17:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.6.19
+
+Sun, 31 Oct 2021 07:17:45 GMT
+
+### Patches
+
+- Bump @microsoft/eslint-config-fast-dna to v2.0.0 (chhol@microsoft.com)
+
+## 1.6.19
+
+Thu, 28 Oct 2021 07:15:32 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.14.3 (nicholasrice@users.noreply.github.com)
 
 ## 1.6.19